requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
Django==2.0.6
kombu==3.0.37
Pillow==5.1.0
pytz==2018.4
redis==2.10.6
vine==1.1.4
django-uuslug==1.1.8
awesome-slugify==1.6.5
